Use 32 bits counter instead of 16bits to handle high precision encoders for motor control. The 16bits value can easily be computed using a cast to uint16_t, or a mask operation on the 16bits LSB (0xFFFF).

+++ src/encoderR4.h
@@ -8,13 +8,13 @@ class EncoderR4{
         EncoderR4(unsigned int pinA, unsigned int pinB);
 
         bool begin();
-        uint16_t read();
-        int16_t readChange();
+        uint32_t read();
+        int32_t readChange();
     private:
         unsigned int _pinA;
         unsigned int _pinB;
         FspTimer _timer{};
-        uint16_t _lastValue;
+        uint32_t _lastValue;
 };
 
 #endif // _ENCODERR4_H_
